Refactor execution process into separate cli command
