+++ .changeset/nice-cougars-impress.md
@@ -1,0 +1,5 @@
+---
+'@chugsplash/plugins': patch
+---
+
+Refactor execution into separate CLI command
